add rocketmq protocol infer

+++ crates/l7-parser/src/rocketmq/constants.rs
@@ -1,4 +1,4 @@
-pub(super) const ROCKETMQ_HEADER_MIN_SIZE: u32 = 21;
+pub(super) const ROCKETMQ_HEADER_MIN_SIZE: usize = 21;
 pub(super) const ROCKETMQ_SERIALIZE_TYPE_JSON: i8 = 0;
 pub(super) const ROCKETMQ_SERIALIZE_TYPE_ROCKETMQ: i8 = 1;
